Backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anthropic==0.2.10
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
dataclasses-json==0.5.7
Django==4.2.1
django-cors-headers==4.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
langchain==0.0.190
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging==23.1
passlib==1.7.4
pycparser==2.21
pydantic==1.10.8
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
tenacity==8.2.2
tokenizers==0.13.3
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
urllib3==2.0.2
yarl==1.9.2
